third_party/WebKit/LayoutTests/typedcssom/stylevalue-normalization/normalize-numeric.html

<!doctype html>
<meta charset="utf-8">
<title>Numeric normalization tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#normalize-numeric">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<script>
'use strict';

function test_numeric_normalization(property, cssText, expected) {
  assert_style_value_equals(CSSStyleValue.parse(property, cssText), expected);
  assert_style_value_equals(
      newDivWithStyle(property + ': ' + cssText).attributeStyleMap.get(property),
      expected);
}

test(() => {
  test_numeric_normalization('line-height', '3.14', CSS.number(3.14));
}, 'Normalizing a <number> returns a number CSSUnitValue');

test(() => {
  test_numeric_normalization('width', '3.14%', CSS.percent(3.14));
}, 'Normalizing a <percentage> returns a percent CSSUnitValue');

test(() => {
  test_numeric_normalization('width', '3.14px', CSS.px(3.14));
}, 'Normalizing a <dimension> returns a CSSUnitValue with the correct unit');

test(() => {
  test_numeric_normalization('width', '0', CSS.px(0));
}, 'Normalizing a <dimension> with a unitless zero returns 0px');

test(() => {
  test_numeric_normalization('z-index', '0', CSS.number(0));
}, 'Normalizing a <number> with a unitless zero returns 0');

test(() => {
  test_numeric_normalization('width',
      'calc(1px + calc(1px) + calc(1px * 2) + 1%)',
      new CSSMathSum(CSS.px(4), CSS.percent(1)));
}, 'Normalizing a <calc> returns simplified expression');

</script>
